Moved ISettings to server source (selectedBranchName => branchName)
